feat(cnd): send logs to a single  incoming queue
